CLI tests (cli-test-1775936545662)

+++ src/main/java/redis/command/builtin/replication/WaitCommand.java
@@ -55,7 +55,7 @@ public record WaitCommand(
 			futures.add(Map.entry(replica, future));
 		});
 
-		var remaining = (long) timeout.toSeconds();
+		var remaining = timeout.toSeconds();
 
 		for (final var entry : futures) {
 			if (acks.get() >= numberOfReplicas) {
